requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
httpx==0.25.2
pillow==10.1.0
numpy==1.26.2
loguru==0.7.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
torch==2.4.1
torchvision==0.19.1
